+++ Ewallet/app/src/main/res/layout/activity_main2.xml
@@ -6,15 +6,23 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin" tools:context=".MainActivity">
 
-
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="100dp"
+        android:textAppearance="?android:attr/textAppearanceLarge"
+        android:text="Total:"
+        android:id="@+id/tv_total"
+        android:layout_alignParentTop="true"
+        android:layout_alignParentLeft="true"
+        android:layout_alignParentStart="true"
+        android:textSize="40dp" />
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
         android:textAppearance="?android:attr/textAppearanceLarge"
         android:text="Input ID number"
         android:id="@+id/textView"
-        android:layout_alignParentTop="true"
+        android:layout_above="@+id/etidnumber"
         android:layout_centerHorizontal="true" />
 
     <EditText
@@ -23,8 +31,7 @@
         android:inputType="number|phone"
         android:ems="10"
         android:id="@+id/etidnumber"
-        android:layout_marginTop="30dp"
-        android:layout_below="@+id/textView"
+        android:layout_above="@+id/button"
         android:layout_centerHorizontal="true" />
 
     <Button
@@ -34,7 +41,9 @@
         android:paddingLeft="40dp"
         android:paddingRight="40dp"
         android:id="@+id/button"
-        android:layout_centerVertical="true"
-        android:layout_centerHorizontal="true"
-        android:onClick="login"/>
+        android:onClick="login"
+        android:layout_alignParentBottom="true"
+        android:layout_centerHorizontal="true" />
+
+
 </RelativeLayout>
