+++ Ewallet/app/src/main/res/layout/activity_welcome_menu.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent" android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    android:paddingBottom="@dimen/activity_vertical_margin" tools:context="app.ewallet.WelcomeMenu">
+    android:paddingBottom="@dimen/activity_vertical_margin" tools:context="app.ewallet.HomeMenu">
 
     <TextView
         android:layout_width="wrap_content"
