+++ Ewallet/app/src/main/res/menu/menu_welcome_menu.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools" tools:context="app.ewallet.WelcomeMenu">
+    xmlns:tools="http://schemas.android.com/tools" tools:context="app.ewallet.HomeMenu">
     <item android:id="@+id/action_settings" android:title="@string/action_settings"
         android:orderInCategory="100" app:showAsAction="never" />
 </menu>
